rename the composition module
